requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
branca==0.8.2
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
cssselect2==0.8.0
curl-cffi==0.5.10
ddgs==9.8.0
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
filelock==3.19.1
folium==0.20.0
fonttools==4.60.1
fpdf2==2.8.4
freetype-py==2.5.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-search-results==2.4.2
googleapis-common-protos==1.72.0
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.1.4
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
narwhals==2.12.0
numpy==2.0.2
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
primp==0.15.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyparsing==3.2.5
pypdf==6.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
streamlit==1.50.0
streamlit-folium==0.25.3
svglib==1.6.0
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
websockets==15.0.1
xyzservices==2025.10.0
zipp==3.23.0
